+++ lib/ui/messages/MessagesConversationPanel.dart
@@ -737,17 +737,19 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
   }
 
   void _openAttachFileMenu() {
-    Analytics().logSelect(target: 'Attach File');
-    showModalBottomSheet(
-      context: context,
-      isScrollControlled: true,
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.vertical(top: Radius.circular(16.0)),
-      ),
-      clipBehavior: Clip.antiAlias,
-      useSafeArea: true,
-      builder: _buildAttachFilePopup,
-    );
+    if (!_submitting && _editingMessage == null) {
+      Analytics().logSelect(target: 'Attach File');
+      showModalBottomSheet(
+        context: context,
+        isScrollControlled: true,
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.vertical(top: Radius.circular(16.0)),
+        ),
+        clipBehavior: Clip.antiAlias,
+        useSafeArea: true,
+        builder: _buildAttachFilePopup,
+      );
+    }
   }
 
   Widget _buildAttachFilePopup(BuildContext context) {
@@ -921,6 +923,8 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
 
   Widget _buildAttachedFileEntry(BuildContext context, int index, {Message? message}) {
     String? name, extension;
+    bool showProgress = false;
+    bool allowRemove = true;
     GestureTapCallback? onTap;
     String? textStyleKey = 'widget.title.dark.small';
     if (message != null) {
@@ -941,27 +945,29 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
       textStyleKey = 'widget.title.dark.small';
     } else {
       dynamic file = _attachedFiles.elementAt(index);
+      name = _getFileName(file);
       FileType type = _getFileType(file);
+      showProgress = (message == null) && (_uploadingFiles[name] == true);
+      allowRemove = !_uploadingFiles.containsKey(name);
       if (type == FileType.image || type == FileType.video) {
-        return _buildAttachedMediaEntry(context, file, type, inMessage: false);
+        return _buildAttachedMediaEntry(context, file, type, inMessage: false, showProgress: showProgress, allowRemove: allowRemove);
       }
       if (type == FileType.audio) {
-        return _buildAudioAttachment(context, file, type, inMessage: false);
+        return _buildAudioAttachment(context, file, type, inMessage: false, showProgress: showProgress, allowRemove: allowRemove);
       }
       textStyleKey = 'widget.title.small';
       if (_uploadingFiles[name] != true) {
         onTap = () => _removeAttachedFiles([file]);
       }
       if (file is PlatformFile) {
-        name = file.name;
         extension = file.extension;
       }
     }
     return _buildAttachmentContainer(
       onTap: message != null ? onTap : null,
-      onRemove: message == null ? onTap : null,
+      onRemove: allowRemove && message == null ? onTap : null,
       inMessage:  message != null,
-      showProgress: message == null && (_uploadingFiles[name] == true),
+      showProgress: showProgress,
       child: Row(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
@@ -1005,7 +1011,7 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
   }
 
   Widget _buildAudioAttachment(BuildContext context, dynamic file, FileType type,
-      {bool inMessage = true}) {
+      {bool inMessage = true, bool showProgress = false, bool allowRemove = true}) {
     String? url;
     Uint8List? bytes;
     if (file is AudioResult) {
@@ -1013,15 +1019,15 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
     }
     else if (file is FileAttachment) {
       url = file.url;
-      bytes = file.data;
     }
     return SizedBox(
       width: 200,
       child: _buildAttachmentContainer(
         inMessage: inMessage,
+        showProgress: showProgress,
         child: Align(alignment: inMessage ? Alignment.center : Alignment.bottomCenter,
             child: AudioPlayerWidget(url: url, bytes: bytes)),
-        onRemove: inMessage ? null : () => _removeAttachedFiles([file]),
+        onRemove: allowRemove && !inMessage ? () => _removeAttachedFiles([file]) : null,
       ),
     );
   }
@@ -1060,15 +1066,11 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
   }
 
   Widget _buildAttachedMediaEntry(BuildContext context, dynamic file, FileType type,
-      {bool inMessage = true}) {
+      {bool inMessage = true, bool showProgress = false, bool allowRemove = true}) {
     String? path, url;
     Uint8List? data;
     if (file is FileAttachment) {
-      if (file.url != null) {
-        url = file.url;
-      } else if (file.data != null) {
-        data = file.data;
-      }
+      url = file.url;
     } else if (kIsWeb && file is PlatformFile) {
       data = file.bytes;
     } else {
@@ -1119,19 +1121,26 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
         }
       } : null,
       behavior: inMessage ? HitTestBehavior.opaque : null,
-      child: Stack(children: [
-        IgnorePointer(ignoring: inMessage, child: AspectRatio(aspectRatio: 1/1, child: widget)),
-        if (!inMessage)
-          Positioned.fill(child:
-            Align(alignment: Alignment.topRight, child:
-              GestureDetector(onTap: () => _removeAttachedFiles([file]),
-                  child: Padding(
-                    padding: const EdgeInsets.all(4.0),
-                    child: Styles().images.getImage('close-circle'),
-                  ))
-            )
-          )
-      ]),
+      child: Stack(
+        alignment: Alignment.center,
+        children: [
+          IgnorePointer(ignoring: inMessage, child: AspectRatio(aspectRatio: 1/1, child: widget)),
+          if (!inMessage && allowRemove)
+            Positioned.fill(child:
+              Align(alignment: Alignment.topRight, child:
+                GestureDetector(onTap: () => _removeAttachedFiles([file]),
+                    child: Padding(
+                      padding: const EdgeInsets.all(4.0),
+                      child: Styles().images.getImage('close-circle'),
+                    ))
+              )
+            ),
+          Visibility(
+            visible: showProgress,
+            child: CircularProgressIndicator(strokeWidth: 3, valueColor: AlwaysStoppedAnimation<Color?>(Styles().colors.fillColorSecondary),),
+          ),
+        ]
+      ),
     );
   }
 
@@ -1305,7 +1314,8 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
   }
 
   Future<void> _createNewMessage(String messageText) async {
-    if (StringUtils.isNotEmpty(messageText) && _conversationId != null && _currentUserId != null && !_submitting) {
+    if (!_submitting && StringUtils.isNotEmpty(messageText) && _conversationId != null && _currentUserId != null) {
+      _submitting = true;
       FocusScope.of(context).requestFocus(FocusNode());
 
       List<FileContentItemReference>? fileRefs;
@@ -1325,7 +1335,6 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
       );
 
       setState(() {
-        _submitting = true;
         _messages.add(tempMessage);
         Message.sortListByDateSent(_messages);
         _shouldScrollToTarget = _ScrollTarget.bottom; //TBD
@@ -1340,27 +1349,43 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
         fileAttachments: fileAttachments,
       );
 
+      Message? newMessage;
+      List<String>? fileIds;
+      if (CollectionUtils.isNotEmpty(newMessages)) {
+        newMessage = newMessages!.first;
+        fileIds = newMessage.fileAttachments?.where((e) => e.type != FileType.file).map((e) => e.id).whereNotNull().toList();
+      }
+      if (CollectionUtils.isNotEmpty(fileIds)) {
+        List<FileContentItemReference>? fileRefs = await Content().getFileContentDownloadUrls(fileIds!, Content.conversationsContentCategory, entityId: _conversationId);
+        for (FileContentItemReference ref in fileRefs ?? []) {
+          for (FileAttachment file in newMessage?.fileAttachments ?? []) {
+            if (ref.id == file.id) {
+              file.url = ref.url;
+            }
+          }
+        }
+      }
+
       if (mounted) {
-        if (newMessages != null && newMessages.isNotEmpty) {
+        if (newMessage != null) {
           setState(() {
-            Message serverMessage = Message.fromOther(newMessages.first, fileAttachments: fileAttachments);
+            Message serverMessage = newMessage!;
             // Update the temporary message with the server's message if needed
             int index = _messages.indexOf(tempMessage);
             if (index >= 0) {
               _messages[index] = serverMessage;
               Message.sortListByDateSent(_messages);
             }
-            _submitting = false;
           });
         } else {
           // If creation failed
           setState(() {
             _messages.remove(tempMessage);
-            _submitting = false;
           });
           AppToast.showMessage(Localization().getStringEx('', 'Failed to send message'));
         }
       }
+      _submitting = false;
     }
   }
 
@@ -1393,20 +1418,17 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
   }
 
   Future<void> _updateEditingMessage(String newText) async {
-    if (_conversationId != null && _editingMessage?.globalId != null) {
+    if (!_submitting && _conversationId != null && _editingMessage?.globalId != null) {
       if (newText == _editingMessage?.message?.trim()) {
         FocusScope.of(context).unfocus();
         setState(() {
           _editingMessage = null;
-          _submitting = false;
           //_shouldScrollToTarget = _ScrollTarget.bottom;
         });
         _inputController.clear();
       }
       else {
-        setState(() {
-          _submitting = true;
-        });
+        _submitting = true;
 
         // Close the keyboard:
         FocusScope.of(context).unfocus();
@@ -1431,7 +1453,6 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
                 Message.sortListByDateSent(_messages);
 
                 _editingMessage = null;
-                _submitting = false;
                 _inputController.clear();
                 // _shouldScrollToTarget = _ScrollTarget.bottom;
               });
@@ -1439,13 +1460,11 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
               debugPrint('Could not find the old message with globalId: ${_editingMessage?.globalId} to replace.');
             }
           } else {
-            setState(() {
-              _submitting = false;
-            });
             AppToast.showMessage(Localization().getStringEx('', 'Failed to update message'));
           }
         }
       }
+      _submitting = false;
     }
   }
 
@@ -1514,6 +1533,10 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
     }
     setState(() {
       _attachedFiles.removeAll(processed);
+      for (dynamic file in processed) {
+        String? name = _getFileName(file);
+        _uploadingFiles.remove(name);
+      }
     });
   }
 
@@ -1609,7 +1632,7 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
     if (file is XFile) {
       return file.path;
     }
-    else if (file is PlatformFile) {
+    else if (!kIsWeb && file is PlatformFile) {
       return file.path;
     }
     return null;
@@ -1664,7 +1687,7 @@ class _MessagesConversationPanelState extends State<MessagesConversationPanel>
       FileType type = _getFileType(file);
       String? name = _getFileName(file);
       FileContentItemReference ref = fileRefs.firstWhere((ref) => ref.name == name, orElse: () => FileContentItemReference());
-      return FileAttachment(name: name, type: type.name, id: ref.id, data: ref.data);
+      return FileAttachment(name: name, type: type.name, id: ref.id);
     }) : [];
   }
 }