+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased - NEOM U
+### Fixed
+- Fix remaining file attachment bugs [#56](https://github.com/rokmetro/neom-app/issues/56)
 
 ## [0.5.7] = 2025-02-14
 ### Fixed